move error creating functions to the error file
